Chore: Refactor some of the prometheus data source types (#107031)
* refactor the types

* more refactoring

* fix import
