+++ packages/grafana-prometheus/src/metric_find_query.ts
@@ -3,6 +3,7 @@ import { map as _map } from 'lodash';
 
 import { MetricFindValue, TimeRange } from '@grafana/data';
 
+import { METRIC_LABEL } from './constants';
 import { PrometheusDatasource } from './datasource';
 import { getPrometheusTime } from './language_utils';
 import {
@@ -13,7 +14,6 @@ import {
   PrometheusQueryResultRegex,
 } from './migrations/variableMigration';
 import { getOriginalMetricName } from './result_transformer';
-import { METRIC_LABEL } from './types';
 
 export class PrometheusMetricFindQuery {
   constructor(
